#98 Same model transition causes corruption
During Finalization of the transition the target resource is unloaded
leading to the deletion of the Capella Project element in the source
resource.

Added a check to avoid unloading the resource when source and target
resources are the same.

Add new tests on all three transitions and dedicated test model.
Tests required a modification of the TRaceabilitySID utility class so
created a new one populating the source and target elements maps
with only elements containing a SID.

Change-Id: I0000000000000000000000000000000000000000
Signed-off-by: Arnaud Dieumegard <arnaud.dieumegard@obeo.fr>